Added waiting animation print, and as such added return carriage to other log statements

+++ twspace.py
@@ -1,4 +1,3 @@
-import os
 import time
 import urllib.request
 from urllib import error
@@ -32,8 +31,10 @@ def send_file(file_path, space_id, twitter_name, space_title, space_date):
         try:
             webhook.send(content=content, file=space_file)
         except discord.HTTPException as e:
+            print("", end="\r")
             logger.error(e.text, exc_info=True)
     else:
+        print("", end="\r")
         logger.error("Could not find space file to send", exc_info=True)
 
 
@@ -54,6 +55,7 @@ def get_m3u8_chunk(base_url, master_url, logger):
 def download(m3u8_id, space_id, twitter_name, space_title, space_date, server):
     logger = create_logger("logfile.log")
     DOWNLOAD_PATH = const.DOWNLOAD
+    SEND_DOWNLOAD = const.SEND_DOWNLOAD
     if DOWNLOAD_PATH == "True":
         DOWNLOAD_PATH = os.getcwd()
     elif not os.path.exists(DOWNLOAD_PATH):
@@ -86,11 +88,14 @@ def download(m3u8_id, space_id, twitter_name, space_title, space_date, server):
             break
         except error.HTTPError as httpError:
             retry += 1
+            print("", end="\r")
             logger.error(httpError, exc_info=True)
             logger.info(f"Retrying(attempt {retry}/{MAX_RETRY}) m3u8 playlist download in {const.SLEEP_TIME} secs...")
             time.sleep(const.SLEEP_TIME)
     if retry == 10:
+        print("", end="\r")
         logger.info(f"Download for {twitter_name}'s {space_id} failed...")
+        return True
     t = t.replace('chunk', base_url+base_addon+m3u8_id+end_chunkurl)
     logger.debug(t)
     filename = f'{space_id}.m3u8'
@@ -102,22 +107,47 @@ def download(m3u8_id, space_id, twitter_name, space_title, space_date, server):
     # Check if the file already exist and if it does remove it
     if os.path.isfile(filename):
         os.remove(filename)
-    # Create a new file with the appropriately replaced chunk url
-    with open(filename, 'w') as f:
-        f.write(t)
+    try:
+        # Create a new file with the appropriately replaced chunk url
+        with open(filename, 'w') as f:
+            f.write(t)
 
-    download_result = subprocess.run(command, capture_output=True, text=True)
-    logger.debug(download_result.stderr)
+        download_result = subprocess.run(command, capture_output=True, text=True)
+        logger.debug(download_result.stderr)
 
-    os.remove(filename)
+        if SEND_DOWNLOAD:
+            send_file(output, space_id, twitter_name, space_title, space_date)
+        print("", end="\r")
+        logger.info(f"Download completed for {space_id}")
+    except Exception:
+        print("", end="\r")
+        logger.error(exc_info=True)
+    finally:
+        # Check if the file already exist and if it does remove it
+        if os.path.isfile(filename):
+            os.remove(filename)
 
-    send_file(output, space_id, twitter_name, space_title, space_date)
-    logger.info(f"Download completed for {space_id}")
+    return True
 
 
 if __name__ == "__main__":
     import twitter_space_bot
+    import threading
+    import os
+
+    def loading_text():
+        loading_string = f"[INFO] Downloading twitter space {space_id} "
+        animation = ["     ", ".    ", "..   ", "...  ", ".... ", "....."]
+        idx = 0
+        while status:
+            print(loading_string + animation[idx % len(animation)], end="\r")
+            time.sleep(0.3)
+            idx += 1
+            if idx == 6:
+                idx = 0
+
     try:
+        status = True
         m3u8_url = input("m3u8 Url: ")
         m3u8_id = twitter_space_bot.get_m3u8_id(m3u8_url)
         server = twitter_space_bot.get_server(m3u8_url)
@@ -125,7 +155,13 @@ if __name__ == "__main__":
         twitter_name = input("twitter name: ")
         space_title = input("space title: ")
         space_date = input("space date: ")
-        print("Download in progress...")
+        t1 = threading.Thread(target=loading_text)
+        t1.start()
         download(m3u8_id, space_id, twitter_name, space_title, space_date, server)
+        status = False
+        exit()
     except Exception as e:
-        print(f"Error encountered...\n{e}")
+        print(f"\rError encountered...\n{e}")
+        while True:
+            input("Exit...")
+            exit()
